Destroy setup forms when purchasing a domain
